Rename Config => config
